pyproject.toml
[tool.poetry]
name = "spcconverter"
version = "0.1.0"
description = ""
authors = ["MoAI522 <urotanre522@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
opencv-python = "^4.6.0"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
